py/pbobj: internalize MICROPY_PY_INSTANCE_ATTRS
This drops the use of the MICROPY_PY_INSTANCE_ATTRS patch.
Instead, offsets are just stored as small ints in the locals_dict, and
we use a custom attribute handler to convert those to objects.

This is made possible by a recent MicroPython patch that allows
to do a generic method lookup if the custom attribute handler
does not find anything.

movehub fw size -160